Refactor RabbitMQ management client and add UserClient
